# openapi.yml
openapi: 3.1.0
info:
  title: CrackCS API
  version: 5.0.0
  description: 세션 인증, 콘텐츠 운영, 근거 검색과 구조화 평가 API
servers:
  - url: /
...
paths:
...
  /api/questions/{questionId}:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    get:
      tags:
        - Public Questions
      operationId: findPublicQuestionById
      summary: 공개 문제 상세 조회
      description: PUBLISHED 상태인 문제만 조회하며, 비공개 문제와 존재하지 않는 문제를 동일하게 처리한다.
      responses:
        '200':
          description: 공개 문제 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicQuestionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/QuestionNotFound'
  /api/admin/knowledge-documents:
    get:
      tags: [Admin Knowledge]
      operationId: findKnowledgeDocuments
      parameters:
        - $ref: '#/components/parameters/TopicIdQuery'
        - $ref: '#/components/parameters/DocumentStatusQuery'
        - $ref: '#/components/parameters/TechnologyVersionQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
      responses:
        '200':
          description: KnowledgeDocument 페이지
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentPageResponse'
    post:
      tags: [Admin Knowledge]
      operationId: createKnowledgeDocument
      summary: 최초 DRAFT 문서 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeDocumentRequest'
      responses:
        '201':
          description: 생성된 문서
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/knowledge-documents/{documentId}:
    parameters:
      - $ref: '#/components/parameters/DocumentIdPath'
    get:
      tags: [Admin Knowledge]
      operationId: findKnowledgeDocumentById
      responses:
        '200':
          description: 문서 상세
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
    patch:
      tags: [Admin Knowledge]
      operationId: updateKnowledgeDocument
      summary: DRAFT 문서 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeDocumentRequest'
      responses:
        '200':
          description: 변경된 문서
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/knowledge-documents/{documentId}/versions:
    parameters:
      - $ref: '#/components/parameters/DocumentIdPath'
    post:
      tags: [Admin Knowledge]
      operationId: createKnowledgeDocumentVersion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KnowledgeDocumentRequest'
      responses:
        '201':
          description: 같은 계열의 새 DRAFT 버전
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/knowledge-documents/{documentId}/review:
    parameters:
      - $ref: '#/components/parameters/DocumentIdPath'
    post:
      tags: [Admin Knowledge]
      operationId: reviewKnowledgeDocument
      responses:
        '200':
          description: 검수 정보가 기록된 문서
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/knowledge-documents/{documentId}/publish:
    parameters:
      - $ref: '#/components/parameters/DocumentIdPath'
    post:
      tags: [Admin Knowledge]
      operationId: publishKnowledgeDocument
      responses:
        '200':
          description: 공개된 문서. 같은 계열의 이전 공개본은 RETIRED가 된다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/knowledge-documents/{documentId}/retire:
    parameters:
      - $ref: '#/components/parameters/DocumentIdPath'
    post:
      tags: [Admin Knowledge]
      operationId: retireKnowledgeDocument
      responses:
        '200':
          description: RETIRED 문서
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/knowledge-documents/{documentId}/chunks:
    parameters:
      - $ref: '#/components/parameters/DocumentIdPath'
    post:
      tags: [Admin Knowledge]
      operationId: generateKnowledgeChunks
      summary: 공개 문서의 검색 Chunk 생성 또는 기존 생성 결과 재사용
      responses:
        '200':
          description: Chunk 생성 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChunkGenerationResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
    get:
      tags: [Admin Knowledge]
      operationId: findKnowledgeChunks
      summary: 문서의 Chunk와 검색 상태 조회
      responses:
        '200':
          description: 원문 순서의 Chunk 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/KnowledgeChunkResponse'
...
  /api/admin/questions:
    get:
      tags:
        - Admin Questions
      operationId: findQuestions
      summary: 문제 목록 조회
      description: 조건에 맞는 문제를 페이지 단위로 조회한다.
      parameters:
        - $ref: '#/components/parameters/TopicIdQuery'
        - $ref: '#/components/parameters/StatusQuery'
        - $ref: '#/components/parameters/DifficultyQuery'
        - $ref: '#/components/parameters/OriginQuery'
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/SizeQuery'
        - $ref: '#/components/parameters/SortQuery'
      responses:
        '200':
          description: 문제 목록 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionPageResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
        - Admin Questions
      operationId: createQuestion
      summary: 초안 문제 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionCreateRequest'
      responses:
        '201':
          description: 문제 등록 성공
          headers:
            Location:
              description: 생성된 문제의 URI
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/admin/questions/{questionId}:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    get:
      tags:
        - Admin Questions
      operationId: findQuestionById
      summary: 문제 상세 조회
      responses:
        '200':
          description: 문제 상세 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/QuestionNotFound'
    patch:
      tags:
        - Admin Questions
      operationId: updateQuestion
      summary: 초안 문제 수정
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionUpdateRequest'
      responses:
        '200':
          description: 문제 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/QuestionNotFound'
  /api/admin/questions/{questionId}/concepts:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    put:
      tags: [Admin Questions]
      operationId: replaceQuestionConcepts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionConceptReplaceRequest'
      responses:
        '200':
          description: 평가 Concept이 교체된 DRAFT 문제
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/questions/{questionId}/review:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    post:
      tags: [Admin Questions]
      operationId: reviewQuestion
      responses:
        '200':
          description: 검수 정보가 기록된 문제
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/questions/{questionId}/publish:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    post:
      tags: [Admin Questions]
      operationId: publishQuestion
      responses:
        '200':
          description: 공개된 문제
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/questions/{questionId}/retire:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    post:
      tags: [Admin Questions]
      operationId: retireQuestion
      responses:
        '200':
          description: RETIRED 문제
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
  /api/admin/questions/{questionId}/versions:
    parameters:
      - $ref: '#/components/parameters/QuestionIdPath'
    post:
      tags: [Admin Questions]
      operationId: createQuestionVersion
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuestionVersionRequest'
      responses:
        '201':
          description: 평가 Concept을 복사한 새 DRAFT 문제 버전
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionResponse'
        '409':
          $ref: '#/components/responses/ContentConflict'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: CRACKCS_SESSION
  parameters:
...
    QuestionIdPath:
      name: questionId
      in: path
      required: true
      description: 문제 ID
      schema:
        type: integer
        format: int64
        minimum: 1
...
    DocumentIdPath:
      name: documentId
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    TopicIdQuery:
      name: topicId
      in: query
      required: false
      description: Topic ID 필터
      schema:
        type: integer
        format: int64
        minimum: 1
...
    DocumentStatusQuery:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/KnowledgeDocumentStatus'
    TechnologyVersionQuery:
      name: technologyVersion
      in: query
      required: false
      schema:
        type: string
        maxLength: 100
    StatusQuery:
      name: status
      in: query
      required: false
      description: 문제 상태 필터
      schema:
        $ref: '#/components/schemas/QuestionStatus'
    DifficultyQuery:
      name: difficulty
      in: query
      required: false
      description: 난이도 필터
      schema:
        $ref: '#/components/schemas/QuestionDifficulty'
    OriginQuery:
      name: origin
      in: query
      required: false
      description: 출처 필터
      schema:
        $ref: '#/components/schemas/QuestionOrigin'
    PageQuery:
      name: page
      in: query
      required: false
      description: 0부터 시작하는 페이지 번호
      schema:
        type: integer
        minimum: 0
        default: 0
    SizeQuery:
      name: size
      in: query
      required: false
      description: 페이지 크기
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortQuery:
      name: sort
      in: query
      required: false
      description: '정렬 조건. 허용 필드: id, topicId, origin, difficulty, status, createdAt, updatedAt. 형식: property,(asc|desc). 여러 조건은 반복해서 전달한다.'
      schema:
        type: array
        items:
          type: string
        default:
          - id,asc
      style: form
      explode: true
...
  responses:
    BadRequest:
      description: 요청 형식 또는 검증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    QuestionNotFound:
      description: 해당 ID의 문제가 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            code: QUESTION_NOT_FOUND
            message: 'Question not found: 999'
            fieldErrors: []
            requestId: 1e85b909-2114-47be-a1c3-1fa47a4a7235
...
    ContentConflict:
      description: 중복 코드·원문 또는 현재 상태에서 허용되지 않는 전이
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
...
    KnowledgeDocumentRequest:
      type: object
      required: [topicId, title, sourceType, content]
      properties:
        topicId:
          type: integer
          format: int64
          minimum: 1
        title:
          type: string
          maxLength: 255
        sourceType:
          $ref: '#/components/schemas/KnowledgeSourceType'
        sourceUrl:
          type: [string, 'null']
          format: uri
          maxLength: 1000
        technologyVersion:
          type: [string, 'null']
          maxLength: 100
        licenseNote:
          type: [string, 'null']
          maxLength: 500
        content:
          type: string
          minLength: 1
    KnowledgeDocumentResponse:
      type: object
      required: [id, topicId, createdByMemberId, title, sourceType, versionSeriesId, documentVersion, content, checksum, status, createdAt, updatedAt]
      properties:
        id: { type: integer, format: int64 }
        topicId: { type: integer, format: int64 }
        createdByMemberId: { type: integer, format: int64 }
        reviewedByMemberId: { type: [integer, 'null'], format: int64 }
        title: { type: string }
        sourceType: { $ref: '#/components/schemas/KnowledgeSourceType' }
        sourceUrl: { type: [string, 'null'] }
        versionSeriesId: { type: string, format: uuid }
        documentVersion: { type: integer, minimum: 1 }
        technologyVersion: { type: [string, 'null'] }
        licenseNote: { type: [string, 'null'] }
        content: { type: string }
        checksum: { type: string, pattern: '^[0-9a-f]{64}$' }
        status: { $ref: '#/components/schemas/KnowledgeDocumentStatus' }
        reviewedAt: { type: [string, 'null'] }
        createdAt: { type: string }
        updatedAt: { type: string }
    KnowledgeDocumentPageResponse:
      allOf:
        - $ref: '#/components/schemas/PageMetadata'
        - type: object
          required: [content]
          properties:
            content:
              type: array
              items:
                $ref: '#/components/schemas/KnowledgeDocumentResponse'
    KnowledgeChunkResponse:
      type: object
      required: [id, sequenceNo, startOffset, endOffset, content, checksum, generationKey, searchStatus]
      properties:
        id: { type: integer, format: int64 }
        sequenceNo: { type: integer, minimum: 0 }
        startOffset: { type: integer, minimum: 0 }
        endOffset: { type: integer, minimum: 1 }
        content: { type: string }
        checksum: { type: string }
        generationKey: { type: string }
        searchStatus:
          type: string
          enum: [KEYWORD_SEARCHABLE, EMBEDDING_PENDING, EMBEDDING_FAILED, READY]
    ChunkGenerationResponse:
      type: object
      required: [generationKey, reused, chunks]
      properties:
        generationKey: { type: string }
        reused: { type: boolean }
        chunks:
          type: array
          items: { $ref: '#/components/schemas/KnowledgeChunkResponse' }
...
    PageMetadata:
      type: object
      required: [page, size, totalElements, totalPages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1, maximum: 100 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    PublicTopicResponse:
      type: object
      required:
        - id
        - code
        - name
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        name:
          type: string
    PublicQuestionResponse:
      type: object
      required:
        - id
        - topic
        - difficulty
        - content
      properties:
        id:
          type: integer
          format: int64
        topic:
          $ref: '#/components/schemas/PublicTopicResponse'
        difficulty:
          $ref: '#/components/schemas/QuestionDifficulty'
        content:
          type: string
      description: 모범 답안과 Concept 평가 정보는 공개 응답에 포함하지 않는다.
...
    QuestionCreateRequest:
      type: object
      required:
        - topicId
        - difficulty
        - content
        - referenceAnswer
      properties:
        topicId:
          type: integer
          format: int64
          minimum: 1
        difficulty:
          $ref: '#/components/schemas/QuestionDifficulty'
        content:
          type: string
          minLength: 1
        referenceAnswer:
          type: string
          minLength: 1
      example:
        topicId: 1
        difficulty: BASIC
        content: 프로세스와 스레드의 차이를 설명하세요.
        referenceAnswer: 프로세스는 자원을 독립적으로 소유하고 스레드는 프로세스 자원을 공유합니다.
    QuestionUpdateRequest:
      type: object
      required:
        - topicId
        - difficulty
        - content
        - referenceAnswer
      properties:
        topicId:
          type: integer
          format: int64
          minimum: 1
        difficulty:
          $ref: '#/components/schemas/QuestionDifficulty'
        content:
          type: string
          minLength: 1
        referenceAnswer:
          type: string
          minLength: 1
    QuestionVersionRequest:
      type: object
      required: [difficulty, content, referenceAnswer]
      properties:
        difficulty:
          $ref: '#/components/schemas/QuestionDifficulty'
        content:
          type: string
          minLength: 1
        referenceAnswer:
          type: string
          minLength: 1
    QuestionConceptReplaceRequest:
      type: object
      required: [concepts]
      properties:
        concepts:
          type: array
          items:
            type: object
            required: [conceptId, weight, required]
            properties:
              conceptId:
                type: integer
                format: int64
                minimum: 1
              weight:
                type: number
                exclusiveMinimum: 0
                maximum: 1
                multipleOf: 0.01
              required:
                type: boolean
    QuestionConceptResponse:
      type: object
      required: [conceptId, code, name, weight, required]
      properties:
        conceptId: { type: integer, format: int64 }
        code: { type: string }
        name: { type: string }
        weight: { type: number }
        required: { type: boolean }
    QuestionResponse:
      type: object
      required:
        - id
        - topicId
        - origin
        - type
        - difficulty
        - content
        - referenceAnswer
        - status
        - versionSeriesId
        - questionVersion
        - concepts
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        topicId:
          type: integer
          format: int64
        origin:
          $ref: '#/components/schemas/QuestionOrigin'
        type:
          $ref: '#/components/schemas/QuestionType'
        difficulty:
          $ref: '#/components/schemas/QuestionDifficulty'
        content:
          type: string
        referenceAnswer:
          type: string
        status:
          $ref: '#/components/schemas/QuestionStatus'
        versionSeriesId:
          type: string
          format: uuid
        questionVersion:
          type: integer
          minimum: 1
        createdByMemberId:
          type: [integer, 'null']
          format: int64
        reviewedByMemberId:
          type: [integer, 'null']
          format: int64
        reviewedAt:
          type: [string, 'null']
        concepts:
          type: array
          items:
            $ref: '#/components/schemas/QuestionConceptResponse'
        createdAt:
          type: string
          description: 서버 로컬 시간 기준 ISO-8601 날짜와 시간
          example: '2026-08-30T22:50:00.123456'
        updatedAt:
          type: string
          description: 서버 로컬 시간 기준 ISO-8601 날짜와 시간
          example: '2026-08-30T22:50:00.123456'
    QuestionPageResponse:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/QuestionSummaryResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        totalElements:
          type: integer
          format: int64
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    ApiErrorResponse:
      type: object
      required:
        - code
        - message
        - fieldErrors
        - requestId
      properties:
        code:
          type: string
        message:
          type: string
        fieldErrors:
          type: array
          items:
            $ref: '#/components/schemas/FieldErrorResponse'
        requestId:
          type: string
          format: uuid
    FieldErrorResponse:
      type: object
      required:
        - field
        - reason
      properties:
        field:
          type: string
        reason:
          type: string
    QuestionDifficulty:
      type: string
      enum:
        - BASIC
        - INTERMEDIATE
        - ADVANCED
    QuestionOrigin:
      type: string
      enum:
        - ADMIN
        - SYSTEM_FOLLOW_UP
    QuestionType:
      type: string
      enum:
        - NORMAL
        - FOLLOW_UP
    QuestionStatus:
      type: string
      enum:
        - DRAFT
        - PUBLISHED
        - RETIRED
    KnowledgeSourceType:
      type: string
      enum: [OFFICIAL_SPEC, OFFICIAL_DOC, INTERNAL_SUMMARY]
    KnowledgeDocumentStatus:
      type: string
      enum: [DRAFT, PUBLISHED, RETIRED]
security:
  - sessionCookie: []